Move services and libs out of app/models/
