+++ app/services/pdf_writer.rb
